requirements.txt
flask>=3.0.0
tensorflow>=2.12.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
keras-tuner>=1.4.7
Pillow>=9.5.0
seaborn>=0.12.2
jupyterlab>=4.0.0
notebook>=7.0.0
pytest>=7.4.0
